Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: Changed some parts of the view for the Main, Bucket, and Workflow and the Model for Bucket and Workflow. Added a button to TaskManager due to null network errors. Added network requests to the Workflow and Bucket presenters and added ResponseObservers for the presenters.
Commit: a3738ae7e92610abedcffd2299841950111aac59