Major changes to how BucketView looks, feels, and functions.
A couple of small changes to make all List<T> into ArrayList<T>, layout changes/preferences, and made the title of the bucket easily changeable in the future by being able to call "setTitle" to change the title label. Removed the "add new task" button because Pollice said no to it. Functionality for it is commented out at the moment. Changed initializations to happen at the conception of the class rather than in the constructor, makes minor difference.

Change-Id: I44ab806c02d1ea6d3c4f96c3d14a9f149d3bdf87