Update patch set 1
Patch Set 1: Code-Review+2

When I added a user with a space in the name the history displayed "John Crespi" as "JohnC".  I am not sure if this is the desired functionality.  other than that it runs.  it might be better to fix this or confirm that it is the desired functionality before you merge.

Patch-set: 1
Label: Code-Review=+2