Update patch set 2
Patch Set 2: Code-Review-1 Verified-1

You seem to have removed the set status from the task view.  Do we still want to be able to change it from the task view or only via drag and drop?

That is the reason I have given this a low score, apart form that the code looks good

Patch-set: 2