Update patch set 2
Patch Set 2: Code-Review+1 Verified+1

Code looks good and it runs.  the height of the colour combo box should be the same as the height for the other combo boxes.  The colour of the text is a little hard to read on the darker colours.  Apart from that it is there and it works

Patch-set: 2
Reviewer: John  Crespi <1002789@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1
Label: Verified=+1