Update patch set 1
Patch Set 1: Code-Review-2

(5 comments)

Make sure you don't accidently submit ~100 lines of changes from reformatting other people's code.  This basically guarantees that you will have a merge conflict with everything else merged into dev.  These conflicts are usually the easiest to fix, but it has to be done over and over again, and I think we've had problems with a bunch of useless merge conflicts hiding the fact that there was an actual conflict to resolve before.

Hint: you can select a block of code in Eclipse and hit Ctrl+Shift+F to only format that block, leaving the rest of the file alone.

Patch-set: 1