Major Changes to how Workflow and Buckets are created and saved. Also allows for changing buckets names to be stored in the DB.
IMPORTANT >>>>>THIS CHANGE WILL REQUIRE A CLEAN DB.<<<<<
Also, because of the new way we're adding buckets to the DB, the first time it's run it will be slow. Without the Thread.sleep(1second) there are thread concurrency problems where certain buckets get added before others, and that's bad m'kay (You'll get completed as the first bucket and new as the either 2nd or 3rd. It's a completely random order each time I ran it without thread.sleeps).

Change-Id: I55c8ce16b89101b1ad1a3d8237c38cbdf3e7d927