Add a servlet for sending email messages from the server
The WPISuite server can now send email notifications to users when requested
from the client through the standard REST API. A field to configure a
per-project Gmail account to send from was also added.

Anything in Janeway can send an email notification just by specifying a message
and a list of users, without either having a hardcoded password in client code
(only the admin might know the password) or a locally running MTA.

There's nothing in the module to let you use it right now, so the only way
to test it is with cURL. I'd like people to review this, but I won't submit it
until there's client support for it.

Change-Id: I3dcbc80d5dea6fddcf15c480369b56047b745db9