Update patch set 1
Patch Set 1:

Problem with adding another button is that we have limited width to work with. We have to either remove the label that tells you what you're doing, or we have to make the buttons Icon only, with no text. I can play around with making it icon only.

Patch-set: 1
Label: Code-Review=0