Update patch set 4
Patch Set 4: Verified-1

Polling works, but when I added a user to a task, the miniview that I was editing on moved to far low in the bucket. The problem was fixed when I reset the view (by toggling archive). On the bright side, the user was added to the other Janeway.

Patch-set: 4
Label: Verified=-1